Fix some warnings

+++ game/src/GameInMap.h
@@ -10,7 +10,7 @@ void OpenTheMap(bool& areSettingsOpen, MapS& map, AllTextures textures);
 void ChooseCountry(bool& areSettingsOpen, int& money, MapS& map,AllTextures textures, Font font, LockedCountries& lockedCountries);
 
 
-void DrawCountry(const char* name, float x, float y, float lengthX, float lenghtY, int countryNum, MapS& map,
+void DrawCountry(const char* name, int x, int y, int lengthX, int lenghtY, int countryNum, MapS& map,
     bool& areSettingsOpen, int& money, bool& lockedCountry, AllTextures textures, Font font);
