+++ game/src/OtherFunctions.cpp
@@ -12,7 +12,7 @@ bool IsMouseInRange(int x1, int x2, int y1, int y2)
     }
 }
 
-Vector2 VecPos(int x, int y)
+Vector2 VecPos(float x, float y)
 {
     Vector2 pos = {0,0};
     pos.x = x;
